+++ crates/cli/src/app.rs
@@ -68,6 +68,14 @@ fn create_styles() -> Styles {
 )]
 pub struct App {
     #[arg(
+        long,
+        global = true,
+        env = "PROTO_DUMP",
+        help = "Dump a trace profile to the working directory"
+    )]
+    pub dump: bool,
+
+    #[arg(
         value_enum,
         long,
         global = true,