+++ crates/core/src/checksum/sha256.rs
@@ -15,7 +15,7 @@ pub fn hash_file_contents<P: AsRef<Path>>(path: P) -> miette::Result<String> {
 
     io::copy(&mut file, &mut sha).map_err(|error| FsError::Read {
         path: path.to_path_buf(),
-        error,
+        error: Box::new(error),
     })?;
 
     let hash = format!("{:x}", sha.finalize());
@@ -25,6 +25,7 @@ pub fn hash_file_contents<P: AsRef<Path>>(path: P) -> miette::Result<String> {
     Ok(hash)
 }
 
+#[tracing::instrument(name = "sha256")]
 pub fn verify_checksum(download_file: &Path, checksum_file: &Path) -> miette::Result<bool> {
     let checksum_hash = hash_file_contents(download_file)?;
     let download_file_name = fs::file_name(download_file);